refactor(net/rdr3): split the player patches into per-subsystem files
