+++ code/components/gta-net-rdr3/include/PlayerLimits.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <StdInc.h>
+
+namespace rage
+{
+static constexpr uint8_t kMaxPlayers = 128;
+}
