+++ code/components/gta-net-rdr3/src/netPedGroupPatch.cpp
@@ -5,6 +5,9 @@
 #include <MinHook.h>
 #include <NetLibrary.h>
 
+#include <PatchUtils.h>
+#include <PlayerLimits.h>
+
 //
 // In RDR3, Ped groups are limited to 68.
 // 32 for players and the rest for peds/posses. For the sake of compatability we want to increase this to 164 (128 players + 36 for peds/posses)
@@ -15,85 +18,8 @@ struct CPedGroup
 	char pad[4832];
 };
 
-static constexpr uint32_t kMaxPedGroups = 128 + 36;
+static constexpr uint32_t kMaxPedGroups = rage::kMaxPlayers + 36;
 static CPedGroup* g_pedGroups;
-
-//TODO: These really need to be moved to some shared file.
-struct PatternPair
-{
-	std::string_view pattern;
-	int offset;
-	int operand_remaining = 4;
-};
-
-static void RelocateRelative(void* base, std::initializer_list<PatternPair> list)
-{
-	void* oldAddress = nullptr;
-	for (auto& entry : list)
-	{
-		auto location = hook::get_pattern<int32_t>(entry.pattern, entry.offset);
-
-		if (!oldAddress)
-		{
-			oldAddress = hook::get_address<void*>(location, 0, entry.operand_remaining);
-		}
-
-		auto curTarget = hook::get_address<void*>(location, 0, entry.operand_remaining);
-		assert(curTarget == oldAddress);
-
-		hook::put<int32_t>(location, (intptr_t)base - (intptr_t)location - entry.operand_remaining);
-	}
-}
-
-struct PatternAbsolutePair
-{
-	std::string_view pattern;
-	int offset;
-	uintptr_t address;
-	int instrLen = 7;
-};
-
-static void RelocateRelativeLocation(std::initializer_list<PatternAbsolutePair> list)
-{
-	uintptr_t oldAddress = 0;
-
-	for (auto& entry : list)
-	{
-	    uint8_t* instructions = reinterpret_cast<uint8_t*>(hook::get_pattern(entry.pattern));
-		uintptr_t instrNext = reinterpret_cast<uintptr_t>(instructions) + entry.instrLen;
-		int32_t newDisp = (int32_t)((int64_t)entry.address- (int64_t)instrNext);
-		hook::put<int32_t>(instructions + entry.offset, newDisp);
-	}
-}
-
-template<int instrLen = 7, int instrOffset = 3>
-static void PatchRelativeLocation(uintptr_t address, uintptr_t newLocation)
-{
-    uint8_t* instructions = reinterpret_cast<uint8_t*>(address);
-	uintptr_t instrNext = address + instrLen;
-	int32_t newDisp = (int32_t)((int64_t)newLocation - (int64_t)instrNext);
-	hook::put<int32_t>(instructions + instrOffset, newDisp);
-}
-
-struct PatternPatchPair
-{
-	std::string_view pattern;
-	int offset;
-	int intendedValue;
-	int newValue;
-};
-
-template<class T>
-static void PatchValue(std::initializer_list<PatternPatchPair> list)
-{
-	for (auto& entry : list)
-	{
-		auto location = hook::pattern(entry.pattern).count(1).get(0).get<T>(entry.offset);
-		auto origVal = *location;
-		assert(origVal == entry.intendedValue);
-		hook::put<T>(location, (T)entry.newValue);
-	}
-}
 
 static HookFunction hookFunction([]()
 {
