+++ code/components/gta-net-rdr3/src/players/PlayerPatches.cpp
@@ -1,0 +1,36 @@
+#include <StdInc.h>
+
+#include <Hooking.h>
+#include <MinHook.h>
+
+#include "PlayerPatches.h"
+
+static HookFunction hookFunction([]()
+{
+	ApplyPlayerDamageArrayPatches();
+	ApplyAnimSceneHandlerPatches();
+	ApplyAnimSceneArrayHandlerPatches();
+	ApplyPlayerCachePatches();
+	ApplyPlayerBandwidthPatches();
+	ApplyPlayerDamageTrackerPatches();
+	ApplyNetObjectPlayerIndexPatches();
+	ApplyNetPlayerMgrPatches();
+	ApplyPlayerIndexComparisons();
+	ApplyPlayerIterationPatches();
+	ApplyPedIntelligencePatches();
+	ApplyPlayerStackResizes();
+	ApplyBubbleJoinPatch();
+	ApplyPlayerBitsetPatches();
+	ApplyDisabledSubsystemPatches();
+	ApplyNetworkObjectMgrPatches();
+
+	MH_Initialize();
+
+	CreateDisabledSubsystemHooks();
+	CreatePlayerFocusHooks();
+	CreatePlayerBitsetHooks();
+	CreatePlayersNearPointHook();
+	CreateDisabledTelemetryHooks();
+
+	MH_EnableHook(MH_ALL_HOOKS);
+});
